Extract serial workers into a module
